Converting manual communicator tests into unit tests
