+++ src/communications/communicator/test/mock_dispatcher.py
@@ -1,0 +1,6 @@
+class MockDispatcher:
+    def __init__(self):
+        self.received = []
+
+    def processEnvelope(self, envelope):
+        self.received.append(envelope)
